move dictionary and words to the same domain
