+++ src/dictionary/word.providers.ts
