+++ src/dictionary/word.ts
